fix: serverless-app
